pyproject.toml
[project]
name = "cognirepo"
version = "0.2.0"
description = "Local cognitive infrastructure layer for AI agents"
readme = "README.md"
requires-python = ">=3.11,<3.15"
license = {text = "AGPL-3.0-or-later"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # core
    "faiss-cpu==1.13.2",
    "numpy==2.4.3",
    "sentence-transformers==5.3.0",
    # MCP server
    "mcp[cli]==1.26.0",
    # REST API
    "fastapi==0.135.2",
    "uvicorn[standard]==0.42.0",
    "python-multipart==0.0.22",
    "starlette==1.0.0",
    # auth
    "PyJWT==2.12.1",
    "bcrypt==5.0.0",
    # HTTP client (CLI --via-api)
    "httpx==0.28.1",
    # graph + indexer
    "networkx>=3.3",
    "watchdog>=4.0",
    "tree-sitter>=0.23",
    # shared
    "pydantic==2.12.5",
    "anyio==4.12.1",
    # gRPC
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "grpcio-health-checking>=1.60",
    # model providers
    "anthropic>=0.50",
    "google-generativeai>=0.8",
    "openai>=1.0",
    # env loading
    "python-dotenv>=1.0",
    # BM25 keyword retrieval (episodic search)
    "rank-bm25>=0.2.2",
]

[project.urls]
Homepage = "https://github.com/ashlesh-t/cognirepo"
"Bug Tracker" = "https://github.com/ashlesh-t/cognirepo/issues"
Documentation = "https://github.com/ashlesh-t/cognirepo/blob/main/ARCHITECTURE.md"
Changelog = "https://github.com/ashlesh-t/cognirepo/blob/main/CHANGELOG.md"
Discord = "https://discord.com/channels/1488386981917360289/1488387271190380636"

[project.optional-dependencies]
languages = [
    "tree-sitter-python>=0.23",
    "tree-sitter-javascript>=0.23",
    "tree-sitter-java>=0.23",
    "tree-sitter-cpp>=0.23",
    "tree-sitter-go>=0.23",
    "tree-sitter-rust>=0.23",
]
cpp = [
    "pybind11>=2.12",
    "scikit-build-core>=0.9",
]
security = [
    "cryptography>=42.0",
    "keyring>=25.0",
]
cli = [
    "rich>=13.7",
    "prompt_toolkit>=3.0",
    "tomli-w>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "pytest-cov>=5.0",
    "httpx>=0.28",
    "prometheus-client>=0.20",
    "build>=1.0",
]

[project.scripts]
cognirepo = "cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["_bm25*", "adapters*", "api*", "cli*", "config*", "cron*", "graph*", "indexer*", "memory*", "orchestrator*", "retrieval*", "rpc*", "security*", "server*", "tools*", "vector_db*"]
exclude = ["venv*", "*.egg-info*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# ── optional C++ build ────────────────────────────────────────────────────────
# Activate by swapping [build-system] to:
#   requires = ["scikit-build-core>=0.9", "pybind11>=2.12"]
#   build-backend = "scikit_build_core.build"
# and running: pip install . --config-settings=cmake.args=-DBUILD_EXT=ON

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "venv/*",
    ".venv/*",
    "rpc/proto/*",
    "scripts/*",
    "deploy/*",
    "docs/*",
    "*.egg-info/*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "except ImportError:",
]
fail_under = 50

[tool.pylint.messages_control]
disable = [
    # unresolvable without optional deps installed
    "import-error",
    "no-member",
    # subprocess check is handled by callers
    "subprocess-run-check",
    # too-few-public-methods is expected for middleware/stub/data classes
    "too-few-public-methods",
    # wrong-import-position is expected for guarded/lazy imports
    "wrong-import-position",
    # import-outside-toplevel is expected for optional/lazy imports throughout
    "import-outside-toplevel",
    # missing docstrings in tests and small helpers are fine
    "missing-function-docstring",
    "missing-class-docstring",
    "missing-module-docstring",
    # useless-object-inheritance: auto-generated proto files use `object`
    "useless-object-inheritance",
    # wrong-import-order: auto-generated proto files have non-standard ordering
    "wrong-import-order",
    # ungrouped-imports: minor style, not a bug
    "ungrouped-imports",
    # duplicate-code: informational only, not actionable in test infrastructure
    "duplicate-code",
]

[tool.pylint.design]
max-locals = 30
max-branches = 25
max-statements = 100
max-returns = 10
max-args = 12
max-positional-arguments = 12
max-attributes = 15
